+++ src/Components/Web.JS/test/EditDistance.test.ts
@@ -1,11 +1,11 @@
 import { expect, test, describe } from '@jest/globals';
-import { compareArrays, ComparisonResult, ItemList, Operation } from '../src/Rendering/DomMerging/EditDistance';
+import { computeEditScript, ComparisonResult, ItemList, Operation } from '../src/Rendering/DomMerging/EditScript';
 
 describe('EditDistance', () => {
   test('should return no operations for empty arrays', () => {
     const before = new ArrayItemList<number>([]);
     const after = new ArrayItemList<number>([]);
-    const result = compareArrays(before, after, exactEqualityComparer);
+    const result = computeEditScript(before, after, exactEqualityComparer);
     expect(result.skipCount).toBe(0);
     expect(result.edits).toBeFalsy();
   });
@@ -13,7 +13,7 @@ describe('EditDistance', () => {
   test('should return insertions when all items are added', () => {
     const before = new ArrayItemList<number>([]);
     const after = new ArrayItemList<number>([1, 2, 3]);
-    const result = compareArrays(before, after, exactEqualityComparer);
+    const result = computeEditScript(before, after, exactEqualityComparer);
     expect(result.skipCount).toBe(0);
     expect(result.edits).toEqual([Operation.Insert, Operation.Insert, Operation.Insert]);
   });
@@ -21,7 +21,7 @@ describe('EditDistance', () => {
   test('should return deletions when all items are removed', () => {
     const before = new ArrayItemList<number>([1, 2, 3]);
     const after = new ArrayItemList<number>([]);
-    const result = compareArrays(before, after, exactEqualityComparer);
+    const result = computeEditScript(before, after, exactEqualityComparer);
     expect(result.skipCount).toBe(0);
     expect(result.edits).toEqual([Operation.Delete, Operation.Delete, Operation.Delete]);
   });
@@ -29,7 +29,7 @@ describe('EditDistance', () => {
   test('should return no edits for identical arrays', () => {
     const before = new ArrayItemList<number>([1, 2, 3]);
     const after = new ArrayItemList<number>([1, 2, 3]);
-    const result = compareArrays(before, after, exactEqualityComparer);
+    const result = computeEditScript(before, after, exactEqualityComparer);
     expect(result.skipCount).toBe(0);
     expect(result.edits).toBeFalsy();
   });
@@ -37,7 +37,7 @@ describe('EditDistance', () => {
   test('should return insert+delete for a replaced item', () => {
     const before = new ArrayItemList<number>([1]);
     const after = new ArrayItemList<number>([2]);
-    const result = compareArrays(before, after, exactEqualityComparer);
+    const result = computeEditScript(before, after, exactEqualityComparer);
     expect(result.skipCount).toBe(0);
     expect(result.edits).toEqual([Operation.Insert, Operation.Delete]);
   });
@@ -45,7 +45,7 @@ describe('EditDistance', () => {
   test('should prefer to substitute rather than insert+delete when allowed', () => {
     const before = new ArrayItemList<number>([1, 2]);
     const after = new ArrayItemList<number>([3, 2]);
-    const result = compareArrays(before, after, (a, b) => (a === b) ? ComparisonResult.Same : ComparisonResult.CanSubstitute);
+    const result = computeEditScript(before, after, (a, b) => (a === b) ? ComparisonResult.Same : ComparisonResult.CanSubstitute);
     expect(result.skipCount).toEqual(0);
     expect(result.edits).toEqual([Operation.Substitute]);
   });
@@ -53,7 +53,7 @@ describe('EditDistance', () => {
   test('should return correct operations for multiple mixed changes', () => {
     const before = new ArrayItemList<number>([1, 2, 3, 4]);
     const after = new ArrayItemList<number>([1, 3, 5, 4]);
-    const result = compareArrays(before, after, exactEqualityComparer);
+    const result = computeEditScript(before, after, exactEqualityComparer);
     expect(result.skipCount).toEqual(1);
     expect(result.edits).toEqual([Operation.Delete, Operation.Keep, Operation.Insert]);
   });
